MyPantry/requirements.txt
django==4.1
pytest-django
django-extensions
cohesion
pytest-cov
#python==3.8
